RavenDB-18204 : modify sharded backup & smuggler tests : - use a base class for shared code - add cluster transactions to InsertData (and remove todos) - change number of expected revisions (we have more docs now since we're using cluster transactions) - assert valid data after import to sharded database
